Window<T> component

+++ src/win/resource_ids.rs
@@ -1,6 +1,6 @@
 pub(super) const IDR_TRAY_MENU: u16 = 101;
 pub(super) const IDI_APP_ICON: u16 = 102;
 
-pub(super) const MENU_SHOW_MAIN_WINDOW: usize = 40001;
-pub(super) const MENU_ABOUT: usize = 40002;
-pub(super) const MENU_EXIT: usize = 40003;
+pub(super) const MENU_SHOW_MAIN_WINDOW: u32 = 40001;
+pub(super) const MENU_ABOUT: u32 = 40002;
+pub(super) const MENU_EXIT: u32 = 40003;
